requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comtypes==1.1.14
discord==2.2.2
discord.py==2.2.2
docopt==0.6.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numpy==1.24.2
pipreqs==0.4.13
PyAudio==0.2.13
pycparser==2.21
pydantic==1.10.7
pyElevenlabs==0.0.2
PyNaCl==1.5.0
python-dotenv==1.0.0
pyttsx3==2.90
requests==2.28.2
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.10.0
tqdm==4.65.0
typing==3.7.4.3
typing_extensions==4.5.0
urllib3==1.26.15
yarg==0.1.9
yarl==1.8.2
